AgriBot/requirements.txt
flask
flask-cors
scikit-learn
tensorflow
deep-translator
pandas
numpy
opencv-python 
Pillow
cohere